tools/imagegen/requirements.txt
diffusers>=0.35.1
transformers>=4.57.0
accelerate>=1.10.1
safetensors>=0.6.2
pillow>=11.3.0
huggingface-hub>=0.34.0
